+++ packages/vendure-plugin-store-credit/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0 (2026-02-03)
+
+- Accept amount as input to the store credit payment handler, to partially pay for an order with store credit
+- Export helper functions to populate wallets for customers
+- BREAKING: DB migration required: Customers can not have multiple wallets with the same name
+
 # 1.0.0 (2026-02-03)
 
 - Initial release